+++ packages/editor-common/web/src/index.js
@@ -87,6 +87,8 @@ export {
   createSelection,
   getBlockType,
   setForceSelection,
+  deleteBlockText,
+  insertString,
 } from './Utils/draftUtils';
 export { isiOS } from './Utils/isiOS';
 export { getSelectionStyles } from './Utils/inlineStyleUtils';
@@ -100,6 +102,8 @@ export {
   DECORATION_MODE,
   PLUGIN_DECORATION_PROPS,
   PLUGIN_DECORATIONS,
+  TEXT_TYPES,
+  CHARACTERS,
 } from './consts';
 
 export {