+++ rollup.externals.js
@@ -11,7 +11,6 @@ const externals = [
   /^@wix\/draft-js$/,
   /^jss$/, //issue with ESM in CJS
   /^jss-preset-default$/, //issue with ESM in CJS
-  /^jss-plugin-nested$/, //issue with ESM in CJS
 ];
 
 const excludedExternalsRegexArr = [