+++ packages/plugin-vertical-embed/web/src/toolbar/insert-buttons.jsx
@@ -5,11 +5,10 @@ import {
 } from 'wix-rich-content-editor-common';
 import { InsertPluginIcon } from '../icons';
 import PostSelectionInputModal from './postSelectionInputModal';
-import { DEFAULTS, VERTICAL_EMBED_TYPES } from '../constants';
+import { DEFAULTS } from '../constants';
 import customStyles from './ModalCustomStyles';
 
 export default ({ helpers, t, settings, isMobile }) => {
-  // TODO: Change icon
   const icon = InsertPluginIcon;
 
   const buttonCreator = (type, tooltipText) => {
@@ -31,9 +30,9 @@ export default ({ helpers, t, settings, isMobile }) => {
     };
   };
 
-  return [
-    buttonCreator(VERTICAL_EMBED_TYPES.PRODUCT, 'Add a product'),
-    buttonCreator(VERTICAL_EMBED_TYPES.EVENT, 'Add an event'),
-    buttonCreator(VERTICAL_EMBED_TYPES.BOOKING, 'Add a booking'),
-  ];
+  const { exposeEmbedButtons = [] } = settings;
+
+  return exposeEmbedButtons.map(verticalType =>
+    buttonCreator(verticalType, `Add a ${verticalType}`)
+  );
 };