+++ CHANGELOG.md
@@ -29,6 +29,9 @@
   - [#903](https://github.com/wix-incubator/rich-content/pull/903) gallery size
 - `editor`
   - [#914](https://github.com/wix-incubator/rich-content/pull/914) updating tooltips keys
+### :house: Internal
+- `wrapper`
+  - [#907](https://github.com/wix-incubator/rich-content/pull/907) converted `wix-rich-content-wrapper` to typescript
 
 
 </details>