+++ CHANGELOG.md
@@ -26,7 +26,7 @@
 - `preview`
   - [#903](https://github.com/wix-incubator/rich-content/pull/903) gallery fixed; resize flapping fixed
 - `gallery`
-  - [#903](https://github.com/wix-incubator/rich-content/pull/903) gallery size
+  - [#909](https://github.com/wix-incubator/rich-content/pull/909) gallery size
 - `editor`
   - [#914](https://github.com/wix-incubator/rich-content/pull/914) updating tooltips keys
 ### :house: Internal