+++ CHANGELOG.md
@@ -19,6 +19,10 @@
 - `editor-common`
   - [#913](https://github.com/wix-incubator/rich-content/pull/913) onChange - calculateDiff is debounced, for better performance
 
+### :house: Internal
+- `wrapper`
+  - [#920](https://github.com/wix-incubator/rich-content/pull/920) refactor: `FullscreenRenderer.tsx`, `ModalRenderer.tsx`
+
 </details>
 <hr/>