+++ CHANGELOG.md
@@ -15,6 +15,18 @@
     Click to see more.
   </summary>
 
+### :bug: Bug Fix
+- `editor-common`
+  - [#913](https://github.com/wix-incubator/rich-content/pull/913) onChange - calculateDiff is debounced, for better performance
+
+### :house: Internal
+- `wrapper`
+  - [#920](https://github.com/wix-incubator/rich-content/pull/920) refactor: `FullscreenRenderer.tsx`, `ModalRenderer.tsx`
+
+</details>
+<hr/>
+
+## 7.3.0 (Apr 13, 2020)
 ### :rocket: New Feature
 - `list`
   - [#815](https://github.com/wix-incubator/rich-content/pull/815) nested lists
@@ -26,13 +38,12 @@
 - `preview`
   - [#903](https://github.com/wix-incubator/rich-content/pull/903) gallery fixed; resize flapping fixed
 - `gallery`
-  - [#903](https://github.com/wix-incubator/rich-content/pull/903) gallery size
+  - [#909](https://github.com/wix-incubator/rich-content/pull/909) gallery size
 - `editor`
   - [#914](https://github.com/wix-incubator/rich-content/pull/914) updating tooltips keys
-
-
-</details>
-<hr/>
+### :house: Internal
+- `wrapper`
+  - [#907](https://github.com/wix-incubator/rich-content/pull/907) converted `wix-rich-content-wrapper` to typescript
 
 ## 7.2.0 (Apr 8, 2020)
 ### :rocket: New Feature